+++ src/models/question/fetch_levels/default.rs
@@ -16,8 +16,8 @@ pub struct DefaultQuestion {
     pub election: Election,
     pub question: MultiLangString,
     pub faculty_code: String,
-    pub student_year_start: i64,
-    pub student_year_end: i64,
+    pub student_year_start: i32,
+    pub student_year_end: i32,
     pub student_program: String,
     pub candidates: Vec<Candidate>,
 }
